Including the node_application_hooks files and the node_main_thread.cpp file in the nmos-cpp-node executable.
It is intended that an application would not include the main.cpp file and just may just call the node_main_thread() function.

In addition, the application would likely replace the node_application_hooks.cpp with one of their own. This would be easier than inserting their changes into node_implementation and having to merge again.