Fix build with boost 1.73.0
Up to boost version 1.72.0, the boost placeholders (_1, _2, etc) were
defined in the global namespace. As of boost 1.73.0, these definitions
are no longer included in the global namespace, so the
boost::placeholders namespace must be explicitly specified.

boost/bind.hpp now has a comment stating:
  For backward compatibility, this header includes
  <boost/bind/bind.hpp> and then imports the placeholders _1, _2,
  _3, ... into the global namespace. Definitions in the global
  namespace are not a good practice and this use is deprecated.
  Please switch to including <boost/bind/bind.hpp> directly,
  adding the using directive locally where appropriate.
  Alternatively, the existing behavior may be preserved by defining
  the macro BOOST_BIND_GLOBAL_PLACEHOLDERS.

As of version 1.73.0, these definitions are no longer in the global
namespace unless BOOST_BIND_GLOBAL_PLACEHOLDERS is defined.

Boost versions older than 1.60.0 define the placeholders directly in the
global namespace, while versions between 1.60.0 and 1.73.0 define the
placeholders in the boost::placeholders namespace, which is then
imported into the global namespace.

To preserve compatibility with versions older than 1.60.0, a new
shim (boost_placeholders) has been defined in the bst:: namespace.

Unfortunately, bst::placeholders is already defined in
slog/all_in_one.h, but is aliased to std::placeholders - which is not
compatible with boost::bind.